+++ crates/stackable-webhook/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- BREAKING: Add support to gracefully shutdown `WebhookServer` and `TlsServer`.
+  Both `WebhookServer::run` and `TlsServer::run` now require passing a shutdown signal, which is any
+  `Future<Output = ()>` ([#1144]).
+
+[#1144]: https://github.com/stackabletech/operator-rs/pull/1144
+
 ## [0.8.1] - 2026-01-07
 
 ### Fixed